feat(deps): move the studio onto @sanity/ui v4 via the catalog
apps/studio and @repo/sanity-schema follow the catalog again, so the whole
workspace is on @sanity/ui@^4.0.0 and there is a single place to bump it.
Adopt v4's subpath entry points for the components they render (Menu,
MenuButton, MenuItem from @sanity/ui/menu, Code from @sanity/ui/code) and
import the static stylesheet at each root entry point.

`sanity@6.9.1` still depends on @sanity/ui@^3.5.1, so the studio bundle
carries both majors until the next studio release. They coexist: styled-
components is a single shared instance, so v4 components pick the theme up
through sanity's v3 ThemeProvider. Box, Flex, Stack, Text, Label, Button,
Code and MenuButton all render unchanged in that tree; only Card reads v4's
own `useRootTheme()` context, which a v3 provider does not supply.
