+++ apps/studio/package.json
@@ -13,7 +13,7 @@
     "@sanity/debug-preview-url-secret-plugin": "^2.0.20",
     "@sanity/icons": "catalog:",
     "@sanity/preview-url-secret": "workspace:^",
-    "@sanity/ui": "^3.4.3",
+    "@sanity/ui": "catalog:",
     "@sanity/vercel-protection-bypass": "^5.0.19",
     "@sanity/vision": "catalog:",
     "@sanity/visual-editing-csm": "workspace:*",
