+++ apps/studio/presentation/CustomHeader.tsx
@@ -1,7 +1,8 @@
 import {CheckmarkIcon} from '@sanity/icons/Checkmark'
 import {CloseIcon} from '@sanity/icons/Close'
 import {EllipsisVerticalIcon} from '@sanity/icons/EllipsisVertical'
-import {Button, Menu, MenuButton, MenuItem} from '@sanity/ui'
+import {Button} from '@sanity/ui'
+import {Menu, MenuButton, MenuItem} from '@sanity/ui/menu'
 import {useState, type FunctionComponent} from 'react'
 import type {PreviewHeaderProps} from 'sanity/presentation'
 import {useSharedState} from 'sanity/presentation'
