+++ apps/studio/presentation/CustomNavigator.tsx
@@ -2,7 +2,8 @@ import {AddIcon} from '@sanity/icons/Add'
 import {BulbOutlineIcon} from '@sanity/icons/BulbOutline'
 import {DocumentIcon} from '@sanity/icons/Document'
 import {SchemaIcon} from '@sanity/icons/Schema'
-import {Box, Button, Card, Flex, Menu, MenuButton, MenuItem, Stack, Text} from '@sanity/ui'
+import {Box, Button, Card, Flex, Stack, Text} from '@sanity/ui'
+import {Menu, MenuButton, MenuItem} from '@sanity/ui/menu'
 import {usePresentationNavigate, usePresentationParams} from 'sanity/presentation'
 import {useIntentLink} from 'sanity/router'
