+++ apps/studio/presentation/DebugStega.tsx
@@ -1,6 +1,7 @@
 import {studioPath, type ContentSourceMap, type ContentSourceMapDocuments} from '@sanity/client/csm'
 import {stegaEncodeSourceMap} from '@sanity/client/stega'
-import {Box, Button, Card, Code, Label, Stack} from '@sanity/ui'
+import {Box, Button, Card, Label, Stack} from '@sanity/ui'
+import {Code} from '@sanity/ui/code'
 import {vercelStegaDecodeAll} from '@vercel/stega'
 import {useMemo} from 'react'
 import {isDocumentSchemaType, type InputProps} from 'sanity'
