+++ apps/studio/sanity.config.ts
@@ -1,3 +1,5 @@
+import '@sanity/ui/styles.css'
+
 import {apiVersion, workspaces} from '@repo/env'
 import {
   crossDatasetReferencesPlugin,
