+++ packages/@repo/sanity-schema/src/page-builder-demo/PageSectionInput.tsx
@@ -1,6 +1,7 @@
 import {EditIcon} from '@sanity/icons/Edit'
 import {RevertIcon} from '@sanity/icons/Revert'
-import {Card, Code, Stack} from '@sanity/ui'
+import {Card, Stack} from '@sanity/ui'
+import {Code} from '@sanity/ui/code'
 import {startTransition, useCallback, useEffect, useMemo, useRef, useState} from 'react'
 import {EMPTY, switchMap} from 'rxjs'
 import {
